Add $preserve_scratch and $always_log so if you want to see details about successful tests, you can.

+++ runtests.sh
@@ -176,8 +176,8 @@ prep_scratch() {
     return 0
 }
 
-discard_scratch() {
-    [ -d "$scratchdir" ] && rm -rf "$scratchdir"
+maybe_discard_scratch() {
+    [ x"$preserve_scratch" != xyes ] && [ -d "$scratchdir" ] && rm -rf "$scratchdir"
     return 0
 }
 
@@ -198,16 +198,23 @@ do
     result=$?
     set -e
 
+    if [ "x$always_log" = xyes -o \( $result != 0 -a $result != 77 -a $result != 78 \) ]
+    then
+	echo "----- $testbase log follows"
+	cat "$scratchdir/test.log"
+	echo "----- $testbase log ends"
+    fi
+
     case $result in
     0)
 	echo "PASS    $testbase"
 	passed=`expr $passed + 1`
-	discard_scratch
+	maybe_discard_scratch
 	;;
     77)
 	echo "SKIP    $testbase"
 	skipped=`expr $skipped + 1`
-	discard_scratch
+	maybe_discard_scratch
 	;;
     78)
         # It failed, but we expected that.  don't dump out error logs, 
@@ -218,9 +225,6 @@ do
 	;;
     *)
 	echo "FAIL    $testbase"
-	echo "----- $testbase failed: log follows"
-	cat "$scratchdir/test.log"
-	echo "----- $testbase log ends"
 	failed=`expr $failed + 1`
 	if [ "x$nopersist" = "xyes" ]
 	then
